remove schedule flyte test class
